+++ app/src/main/java/com/unir/sheet/data/model/CharacterItemDetail.kt
@@ -3,7 +3,7 @@ package com.unir.sheet.data.model
 
 data class CharacterItemDetail(
     val item: Item,
-    val characterId: Int,
+    val characterId: Long,
     val quantity: Int
 
 )
