+++ app/src/main/java/com/unir/sheet/domain/repository/CharacterRepository.kt
@@ -4,7 +4,7 @@ import com.unir.sheet.data.model.CharacterEntity
 
 interface CharacterRepository {
     suspend fun getCharactersByUserId(userId: Int): Result<List<CharacterEntity>>
-    suspend fun getCharacterById(id: Int):  Result<CharacterEntity?>
+    suspend fun getCharacterById(id: Long):  Result<CharacterEntity?>
     suspend fun saveCharacter(character: CharacterEntity): Result<CharacterEntity>
     suspend fun deleteCharacter(character: CharacterEntity): Result<Unit>
