+++ app/src/main/java/com/unir/sheet/domain/repository/SkillRepository.kt
@@ -6,8 +6,8 @@ import com.unir.sheet.data.model.Skill
 
 interface SkillRepository {
     suspend fun getAllSkills() : Result<List<Skill>>
-    suspend fun addSkillToCharacter(characterId: Int, skillId: Int) : Result<Unit>
-    suspend fun deleteSkillFromCharacter(characterId: Int, skillId: Int) : Result<Unit>
-    suspend fun getSkillsFromCharacter(characterId: Int) : Result<List<Skill>>
-    suspend fun addDefaultSkills(characterId: Int, skillIds: List<Int>): Result<Unit>
+    suspend fun addSkillToCharacter(characterId: Long, skillId: Int) : Result<Unit>
+    suspend fun deleteSkillFromCharacter(characterId: Long, skillId: Int) : Result<Unit>
+    suspend fun getSkillsFromCharacter(characterId: Long) : Result<List<Skill>>
+    suspend fun addDefaultSkills(characterId: Long, skillIds: List<Int>): Result<Unit>
 }
