+++ app/src/main/java/com/unir/sheet/ui/screens/character/characterDetail/CharacterDetailScreen.kt
@@ -36,7 +36,7 @@ import com.unir.sheet.ui.viewmodels.CharacterViewModel
 
 @Composable
 fun CharacterDetailScreen(
-    characterId : Int,
+    characterId : Long,
     characterViewModel: CharacterViewModel = LocalCharacterViewModel.current,
 ){
